Refactor image utilities: reorder imports, add `TmpColorType` to `ColorType` conversion, and improve test formatting.
